do not skip horizon jobs
On Nov 18th or thereabouts, I intended to make our horizon
work like upstream in that it would skip some jobs when only
documentation/etc was changed.

- any sort of devstack/dsvm job for horizon was skipped for
doc, releasenotes, unittest changes, etc.

- the rest, non-dsvm tests, were also skipped for .rst and doc/* changes:
^.*horizon-(python27|python34|tox-py27dj.*|npm-run-.*|selenium-headless|dsvm-integration)$

This change 'reverts' this behavior, so we always build everything
This is done because we think that upstream merges may not be passing
the 'is not a doc change' tests.

Change-Id: I9ab44c1765ff7fc1028cbf12d6fa7d89628c26c0