upgrade dib-utils in scenario and build-image jobs
Do the hidden extra initializations that dsvm_int_tests method
does in causes where we don't call this method.

In the end, this will all be moved in some fashion into
trove-integratin - whether that be new modes of dsvm_int_tests,
or a similar method to wrap execution of these new scenario tests.

Change-Id: Iee357feffc68b462976031704950b08e2de7cd54