fix failure in yaml formatting
I must have not noticed something in the checks.
The previous push had a slight formatting error in the yaml
which threw zuul for a loop.
Also include the merge-check template, as it seem required.

I have some priviledges to work out to make this function.
something in zuul/zuul-merger needs to be able to access
private repos.

Change-Id: I2b71ccd323dbf73a37fb3fb83381242b4976d44e